move methods from RepositoryExtensions
